Move cost aggregator from main.py to separate module
