+++ app/main.py
@@ -6,6 +6,7 @@ from pydantic import BaseSettings
 from starlette.websockets import WebSocket
 from starlette.websockets import WebSocketDisconnect
 
+from app.api import cost_aggregator
 from app.api.modules import modules
 
 
@@ -15,6 +16,7 @@ class Settings(BaseSettings):
     These default settings get overridden by environment variables.
     @see https://fastapi.tiangolo.com/advanced/settings/
     """
+
     host: str = "localhost:8000"
 
 
@@ -117,69 +119,7 @@ async def get() -> dict[str, list[dict[str, Any]]]:
     return {"responses": modules.response_schemas}
 
 
-# Register modules
+# Register routers
 modules.include_routers(app)
+app.include_router(cost_aggregator.router)
 # ==================================================================
-# Fucntions to call for each module are hard-coded for now
-# Todo: use interface to obtain function to call for each module
-# Todo: perform input sanity check and handle module errors per module
-# Todo: move into own module
-
-from pydantic import BaseModel, conlist, confloat
-from typing import Any
-
-class CostItem(BaseModel):
-    module: str
-    properties: Any
-
-class CostPath(BaseModel):
-    title: str
-    cost_items: conlist(CostItem, min_items=1)
-
-class CostAggregatorRequest(BaseModel):
-    cost_paths: conlist(CostPath, min_items=1)
-
-class CostItemResponse(BaseModel):
-    cost_item: CostItem
-
-class CostPathResponse(BaseModel):
-    title: str
-    total_carbon_kg: confloat(ge=0.0)
-    cost_items: conlist(CostItemResponse)
-
-class CostAggregatorResponse(BaseModel):
-    cost_paths: conlist(CostPathResponse)
-
-@app.post("/cost_aggregator")
-async def cost_aggregator(request: CostAggregatorRequest) -> CostAggregatorResponse:
-    cost_paths = []
-    for cost_path in request.cost_paths:
-        cost_items = []
-        total_carbon_kg = 0.0
-        for cost_item in cost_path.cost_items:
-            # Looking for the module with this return type
-            for module in module_list:
-                if cost_item.module == module.request()["title"]:
-                    # Calling the code that computes the information
-                    # Todo: extend and use module interface
-                    if cost_item.module == "OnlineDetails":
-                        from app.api.vc_calculator import online_calculator
-                        from vc_calculator.interface import OnlineDetails
-                        res = online_calculator(
-                            OnlineDetails(**cost_item.properties))
-                        import statistics
-                        total_carbon_kg += statistics.mean(
-                            [res.total_emissions.low, res.total_emissions.high])
-                    elif cost_item.module == "FlightCalculatorRequest":
-                        from app.api.flight_calculator import flight_calculator, FlightCalculatorRequest
-                        res = flight_calculator(
-                            FlightCalculatorRequest(**cost_item.properties))
-                        total_carbon_kg += res.total_carbon_kg
-            cost_items.append( CostItemResponse(cost_item = cost_item) )
-        cost_paths.append( CostPathResponse(cost_items = cost_items, title = cost_path.title, total_carbon_kg = total_carbon_kg) )
-    return CostAggregatorResponse(cost_paths = cost_paths)
-    #return {"message": "OK"}
-
-# Include the module routers
-for module in module_list:
-    app.include_router(module.router)
